Move defaultPaletteId to BaseEffect
This is an artifact of the original effect class concept, but I believe
it's no longer needed. So removing it will be easier if it is just one
place.
Even if it turns out to be needed, the vast majority of effects use 0
anyway, so having it as a default without all effects needing to declare
it makes sense.